# docs/debate-api-openapi.yaml
openapi: 3.1.0
info:
  title: Debate Platform MVP API
  version: 1.0.0
servers:
  - url: http://localhost:3000
paths:
  /api/debates:
    get:
      summary: List debates
      parameters:
        - in: query
          name: limit
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
      responses:
        '200':
          description: Debate list
    post:
      summary: Create debate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title, motionText, format, domain]
              properties:
                title: { type: string }
                motionText: { type: string }
                format: { type: string, enum: [short, long] }
                domain: { type: string, enum: [politics, sports, general] }
      responses:
        '201':
          description: Debate created
...